suggestions from code review
